Format code and remove console logs
