+++ src/components/dog-card-large/DogCardLarge.scss
@@ -4,16 +4,13 @@
     padding: 20px;
     border-radius: 8px;
     margin-bottom: 10px;
-
     position: fixed;
     top: 50%;
     left: 50%;
     transform: translate(-50%, -50%);
     z-index: 100;
-    
     display: flex;
     justify-content: space-between;
-
     text-align: left;    
     background-color: #ffffff;
     box-shadow: 11px 11px 23px -11px rgba(0, 0, 0, 0.06);
@@ -106,9 +103,6 @@
             }
         }
 
-        
-
-
         .card-large-gender-icon {
             position: absolute;
             top: 163px;
@@ -120,7 +114,6 @@
             top: 224px;
             right: 0;
         }
-
     }
 
     .card-large-image {
@@ -142,8 +135,6 @@
             display: none;
         }
 
-
-
         .card-large-close-icon {
             width: 33px;
             height: 33px;
@@ -151,7 +142,6 @@
             top: 8px;
             right: 8px;
             cursor: pointer;
-            
         }
         
         .card-large-favourite-icon {
@@ -162,6 +152,5 @@
             bottom: 12px;
             cursor: pointer;
         }
-        
     }
 }
