+++ src/components/dog-card-small/DogCardSmall.jsx
@@ -2,13 +2,10 @@ import { useState, React } from 'react';
 import './DogCardSmall.css';
 import DogCardLarge from '../dog-card-large/DogCardLarge';
 
-
 import favouriteIcon from '../../assets/favourite-filled.svg';
 import expandIcon from '../../assets/expand-solid.svg';
-
 import femaleIcon from '../../assets/female.svg';
 import maleIcon from '../../assets/male.svg';
-
 import smallDogIcon from '../../assets/small-size-dog.png';
 import mediumDogIcon from '../../assets/medium-size-dog.png';
 import largeDogIcon from '../../assets/large-size-dog.png';
@@ -39,7 +36,6 @@ export default function DogCardSmall({dog}) {
     return (
         <>
         <div className='card-small'>
-
             <div className='card-small-details'>
                 <div className='card-small-name'>{dog.name}</div>
                 <div className='card-small-breed-text'>fajtája:</div>
@@ -61,22 +57,18 @@ export default function DogCardSmall({dog}) {
                 
                 <img className='card-small-size-icon' src={dogSizeImage(dog.size)} alt='medium dog' />
             </div>
-
             <div className='card-small-image'>
                 <img className='card-small-dog-image' src={dogImage} alt='dog' />
                 <img className='card-small-favourite-icon' src={favouriteIcon} alt='heart' />
                 <img className='card-small-expand-icon' src={expandIcon} alt='expand' onClick={handleExpand} />
             </div>
-
         </div>
 
-        {
-        isExpanded
+        {isExpanded
         &&
         <div className='dog-card-large-background' onClick={handleExpand}>
             <DogCardLarge dog={dog} handleExpand={handleExpand}/>
-        </div>
-        }
+        </div>}
         </>
     )
 }
