+++ src/components/dog-card-small/DogCardSmall.scss
@@ -4,7 +4,6 @@
     padding: 20px;
     border-radius: 8px;
     margin-bottom: 10px;
-
     display: flex;
     justify-content: space-between;
     background-color: #ffffff;
@@ -64,7 +63,6 @@
             top: 71px;
             right: 0;
         }
-
     }
 
     .card-small-image {
@@ -84,7 +82,6 @@
             top: 8px;
             right: 8px;
             cursor: pointer;
-            
         }
         
         .card-small-favourite-icon {
@@ -95,6 +92,5 @@
             bottom: 4px;
             cursor: pointer;
         }
-        
     }
 }
