+++ src/components/footer/Footer.scss
@@ -1,12 +1,11 @@
 footer {
     width: 100%;
     height: 250px;
-
+    bottom: 0;
+    position: absolute;
     display: flex;
     align-items: center;
     padding: 0 75px;
-
-    
     background-color: #FBFAFE;
     
     .footer-container {
@@ -24,25 +23,28 @@ footer {
             background-size: contain;
             background-repeat: no-repeat;
             cursor: pointer;
-    
         }
     
         .navbar-bottom {
             display: flex;
     
-    
             nav {
-                padding: 0 45px;
                 display: flex;
                 align-items: center;
-                column-gap: 25px;
-    
-                ul {
-                    font-size: 16px;
-                    font-weight: 600;
-                    color: #000000;
-                    cursor: pointer;
                 
+                ul {
+                    padding: 0 45px;
+                    display: flex;
+                    align-items: center;
+                    column-gap: 20px;
+
+                    li {
+                        font-size: 16px;
+                        font-weight: 600;
+                        color: #000000;
+                        cursor: pointer;
+                        list-style-type: none;
+                    }
                 }
             }
         }
@@ -62,5 +64,4 @@ footer {
             }
         }
     }
-
 }
