+++ src/components/header/Header.jsx
@@ -1,32 +1,20 @@
-import React from 'react';
-import { useState } from 'react';
+import { React, useContext } from 'react';
+import { HashLink } from 'react-router-hash-link';
+import { NavLink, useNavigate } from 'react-router-dom';
+
 import './Header.css';
+
 import logoutIcon from '../../assets/logout.svg';
 import profileIcon from '../../assets/profile.svg';
 
 import AuthContext from '../../contexts/AuthContext';
 
-//import AnchorLink from 'react-anchor-link-smooth-scroll'; --- v2 has come out since
-//import AnchorLink from 'react-anchor-link-smooth-scroll-v2';
-
-import { NavLink, useNavigate } from 'react-router-dom';
-// NavLink can be styled based on whether it is active or not
-
-import { HashLink } from 'react-router-hash-link';
-import { useContext } from 'react';
-
-// this function is for Hashlink to scroll with offset of 120px
 const scrollWithOffset = (el) => {
     const yCoordinate = el.getBoundingClientRect().top + window.pageYOffset;
     const yOffset = -120; 
     window.scrollTo({ top: yCoordinate + yOffset, behavior: 'smooth' }); 
 }
-/* 
-inline solution, but with no smooth scroll: 
-scroll={el => { el.scrollIntoView(true); window.scrollBy(0, -120) }} 
-*/
 
-// scroll to bottom when Contact is clicked regardless of page we are on
 function scrollToBottom() {
     window.scroll({
         top: document.body.offsetHeight,
@@ -35,31 +23,25 @@ function scrollToBottom() {
     });
 }
 
-
 export default function Header() {
 
     const navigate = useNavigate();
 
     const { userData, isLoggedIn, logOut } = useContext(AuthContext);
-    //console.log(userData)
 
     const headerLogOut = () => {
         logOut();
         navigate('/login')
     }
 
-
     return (
         <header>
             <div className='header-container'>
-
+                
                 <div className='navbar-top'>
                     <HashLink to='/home'><div className='logo'></div></HashLink>
                     <nav>
                         <ul>
-
-                            {/* <li><AnchorLink offset= '120' href='#random-name-section'>Random</AnchorLink> </li> */}
-
                             <li>
                                 <HashLink 
                                     smooth 
@@ -68,7 +50,6 @@ export default function Header() {
                                     >Főoldal
                                 </HashLink>
                             </li>
-
                             <li>
                                 <HashLink 
                                     smooth 
@@ -77,7 +58,6 @@ export default function Header() {
                                     >Névválasztó
                                 </HashLink>
                             </li>
-
                             <li>
                                 <HashLink 
                                     smooth 
@@ -86,7 +66,6 @@ export default function Header() {
                                     >Random
                                 </HashLink>
                             </li>
-                            
                             {!isLoggedIn
                             &&
                             <li>
@@ -97,17 +76,7 @@ export default function Header() {
                                     >Regisztráció
                                 </HashLink>
                             </li>}
-
-                            {/* <li>
-                                <HashLink 
-                                    smooth 
-                                    scroll={el => scrollWithOffset(el)}
-                                    to='/#footer'
-                                    >Kontakt
-                                </HashLink>
-                            </li> */}
                             <li onClick={scrollToBottom}>Kontakt</li>
-
                         </ul>
                     </nav>
                 </div>
