+++ src/components/header/Header.scss
@@ -1,15 +1,11 @@
 header {
     width: 100%;
     height: 72px;
-
     padding: 0 75px;
-   
     position: fixed;
     top: 0;
     z-index: 1000;
-    
     background-color: #ffffff;
-    
     box-shadow: 0px 1px 7px 0px rgba(0, 0, 0, 0.12);
     
     .header-container {
@@ -20,7 +16,6 @@ header {
         flex-direction: row;
         justify-content: space-between;
         align-items: center;
-
     
         .navbar-top {
             display: flex;
@@ -32,7 +27,6 @@ header {
                 background-size: contain;
                 background-repeat: no-repeat;
                 cursor: pointer;
-    
             }
             
             nav {
@@ -50,7 +44,6 @@ header {
                         color: #F17E5B;
                         cursor: pointer;
                         list-style-type: none;
-                    
                     }
                 }
             }
@@ -92,7 +85,5 @@ header {
                 cursor: pointer;
             }
         }
-
     }    
-
 }
