+++ src/components/hero/Hero.scss
@@ -22,8 +22,6 @@
             text-align: center;
             font-size: 18px;
         }
-
-
     }
 
     .hero-animation {
@@ -31,5 +29,4 @@
         display: flex;
         align-items: center;
     }
-
 }
