+++ src/components/name-card/NameCard.scss
@@ -6,13 +6,10 @@
     margin-right: 21px;
     margin-bottom: 10px;
     margin-top: 10px;
-    
     position: relative;
     display: inline-block;
-    
     background-color: #FFFFFF;
     box-shadow: 11px 11px 23px -11px rgba(0, 0, 0, 0.06);
-
     
     div {
         margin-right: 40px;
@@ -29,5 +26,4 @@
         top: 12px;
         cursor: pointer;
     }
-    
 }
