+++ src/components/registration/Registration.scss
@@ -59,13 +59,11 @@
         
         .register-error-message {
             background-color: #FBC2B1;
-            }
+        }
 
         .register-success {
             background-color: #B1FBBD;
         }
-        
-
 
         button {
             width: 183px;
@@ -81,7 +79,6 @@
     .sitting-jack-russel {
         width: 282px;
         height: 273px;
-
         position: absolute;
         right: 78px;
         bottom: 7px;
