+++ src/components/search/Search.css
@@ -43,7 +43,6 @@
   background-color: #FBFAFE;
   -webkit-box-shadow: 11px 16px 20px -5px rgba(0, 0, 0, 0.09);
           box-shadow: 11px 16px 20px -5px rgba(0, 0, 0, 0.09);
-  /* breed */
 }
 
 #search-section .search-container .search-form-container form fieldset {
@@ -75,46 +74,6 @@
 #search-section .search-container .search-form-container form fieldset .input-hidden:checked + .input-label {
   color: #FFFFFF;
   background-color: #F17E5B;
-}
-
-#search-section .search-container .search-form-container form .breed-dropdown {
-  /*                     
-                    label {
-                        position: relative;
-                        
-                        &::before {
-                            content: '';
-                            width: 42px;
-                            height: 42px;
-                            border-radius: 6px;
-                            position: absolute;
-                            top: -12px;
-                            right: 0;
-                            background-color: #F17E5B;
-                            z-index: -10;
-                        }
-
-                    }    
-                    */
-  /*                     
-                    .css-b62m3t-container {
-                    width: 300px;
-                    margin: 5px 0;
-                    border: 2px solid #F17E5B;
-                    border-radius: 6px;
-
-                        if no item is picked in dropdown
-                        .css-1s2u09g-control {
-                            border: none
-                        }
-
-                        if item is picked in dropdown 
-                        .css-1pahdxg-control {
-                            border: none;
-                            box-shadow: none;
-                        }
-                    }
-                    */
 }
 
 #search-section .search-container .search-form-container form .breed-dropdown select {
@@ -151,18 +110,6 @@
   margin-bottom: 60px;
   position: relative;
   text-align: center;
-  /*         
-        &::before {
-        content: '';
-        max-width: 100%;
-        background-color: aqua;
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        z-index: -1;
-        }
-        */
 }
 
 #search-section .search-results-container .no-result-found {
