+++ src/components/send-name/SendName.jsx
@@ -1,26 +1,10 @@
 import { useState, useEffect, useContext, React } from 'react';
-//import Select from 'react-select';
+import './SendName.css';
 
 import AuthContext from '../../contexts/AuthContext';
 
-
-import './SendName.css';
-
-const options = [
-    { value: 'not defined', label: 'Mindegy' },
-    { value: 'keverék', label: 'Keverék' },
-    { value: 'pumi', label: 'Pumi' },
-    { value: 'tacskó', label: 'Tacskó' },
-    { value: 'vizsla', label: 'Vizsla' },
-    { value: 'border collie', label: 'Border collie' },
-    { value: 'törpe pincser', label: 'Törpe pincser' },
-    { value: 'mudi', label: 'Mudi' },
-    { value: 'törpespicc', label: 'Törpespicc' }
-]
-
 export default function SendName({allDogs}) {
 
-    //console.log(allDogs)
     const { userData } = useContext(AuthContext);
     
     const [dogBreeds, setDogBreeds] = useState([])
@@ -37,18 +21,14 @@ export default function SendName({allDogs}) {
         setDogBreeds(dogBreeds.sort())
     }, [allDogs])
     
-    //console.log(dogBreeds)
-
-
-    // - - - -  input change handler - - - - 
+    // - - - -  input change handler: radio, select - - - - 
     const handleChange = (event) => {
         const name = event.target.name;
         const value = event.target.value;
         setInputs(values => ({...values, [name]: value}))
     }
 
-    // - - - -  input change handler - Checkboxes - - - - 
-
+    // - - - -  input change handler: checkboxes - - - - 
     const handleChangeCheckbox = (event) => {
         const name = event.target.name;
         const newTrait = event.target.value;
@@ -64,16 +44,14 @@ export default function SendName({allDogs}) {
            //console.log(newTraits)
             setInputs((values) => ({...values, [name]: newTraits}))
         }
-
     }
 
-    // - - - - input change handler - File - - - - 
+    // - - - - input change handler: file upload - - - - 
     const handleChangeFile = (event) => {
         setSelectedFile(event.target.files[0])
-        
     }
     
-    // - - - - handle Submit - - - - 
+    // - - - - handle form submit - - - - 
     const handleSubmit = async function(event) {
         event.preventDefault();
 
@@ -81,21 +59,14 @@ export default function SendName({allDogs}) {
         formData.append('file', selectedFile);
         formData.append('object', JSON.stringify(inputs))
 
-        
         await fetch('https://doggobase-api.onrender.com/addnewdog', {
             method: "POST",
-            //headers: { 'Content-Type': 'multipart/form-data' },
             body: formData
         })
-        
-        console.log("posted")
-
     }
 
     return (
-        
         <form>
-
             <fieldset>
                 <label htmlFor='name-of-dog'>Név:</label>
                 <input 
@@ -108,10 +79,8 @@ export default function SendName({allDogs}) {
                     onChange={handleChange}
                 />
             </fieldset>
-
             <fieldset className='gender-radio'>
                 <legend>Neme:</legend>
-
                 <input 
                     className='input-hidden'
                     type='radio' 
@@ -122,7 +91,6 @@ export default function SendName({allDogs}) {
                     checked={inputs.gender === 'fiú'}
                 />
                 <label className='input-label' htmlFor="male">Fiú</label>
-
                 <input 
                     className='input-hidden' 
                     type='radio' 
@@ -133,12 +101,9 @@ export default function SendName({allDogs}) {
                     checked={inputs.gender === 'lány'}
                     />
                 <label className='input-label' htmlFor="female">Lány</label>
-
             </fieldset>
-
             <fieldset className='size-radio'>
                 <legend>Mérete:</legend>
-
                 <input 
                     className='input-hidden' 
                     type='radio' 
@@ -149,7 +114,6 @@ export default function SendName({allDogs}) {
                     checked={inputs.size === 'kicsi'}
                 />
                 <label className='input-label' htmlFor="small">Kicsi</label>
-
                 <input 
                     className='input-hidden' 
                     type='radio' 
@@ -160,7 +124,6 @@ export default function SendName({allDogs}) {
                     checked={inputs.size === 'közepes'}
                 />
                 <label className='input-label' htmlFor="medium">Közepes</label>
-
                 <input 
                     className='input-hidden' 
                     type='radio' 
@@ -172,7 +135,6 @@ export default function SendName({allDogs}) {
                 />
                 <label className='input-label' htmlFor="large">Nagy</label>
             </fieldset>
-
             <fieldset className='breed-dropdown'>
                 <legend>Fajtája:</legend>
                 <label htmlFor="breed">
@@ -189,10 +151,8 @@ export default function SendName({allDogs}) {
                     </select>
                 </label>
             </fieldset>
-
             <fieldset className='traits-checkbox'>
                 <legend>Egyéb jellemzők (opcionális):</legend>
-
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -202,7 +162,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                 />
                 <label className='input-label' htmlFor='smart' >Okos</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -212,7 +171,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                 />
                 <label className='input-label' htmlFor='playful'>Játékos</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -222,7 +180,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                     />
                 <label className='input-label' htmlFor='fluffy'>Bundás</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -232,7 +189,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                     />
                 <label className='input-label' htmlFor='shy'>Félénk</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -242,7 +198,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                     />
                 <label className='input-label' htmlFor='active'>Aktív</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -252,7 +207,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                 />
                 <label className='input-label' htmlFor='hungry'>Falánk</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -262,7 +216,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                 />
                 <label className='input-label' htmlFor='barking'>Ugatós</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -272,7 +225,6 @@ export default function SendName({allDogs}) {
                     onChange={handleChangeCheckbox}
                 />
                 <label className='input-label' htmlFor='cuddly'>Bújós</label>
-                
                 <input 
                     className='input-hidden' 
                     type='checkbox' 
@@ -283,7 +235,6 @@ export default function SendName({allDogs}) {
                 />
                 <label className='input-label' htmlFor='lazy'>Lusta</label>
             </fieldset>
-
             <label htmlFor='image-upload-input'>Kép feltöltése:</label>
             <input 
                 type='file'
@@ -292,11 +243,7 @@ export default function SendName({allDogs}) {
                 onChange={handleChangeFile}
                 required={true}
             />
-
             <button onClick={handleSubmit}>Beküldöm a kutyát!</button>
-
         </form>
-        
-
     )
 }
