+++ src/pages/Layout.jsx
@@ -2,13 +2,12 @@ import React from 'react';
 import Header from '../components/header/Header';
 import Footer from '../components/footer/Footer';
 
-
 export default function Layout({children}) {
     return (
         <>
-            <Header />
-            <main>{children}</main>
-            <Footer />
+        <Header />
+        <main>{children}</main>
+        <Footer />
         </>
     )
 }
