+++ src/pages/nopage/NoPage.css
@@ -1,0 +1,33 @@
+.no-page-image-container {
+  width: calc(100vw + 75px);
+  height: 100vh;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  position: fixed;
+  left: -75px;
+  background-color: #C9E2EB;
+}
+
+.no-page-image-container p {
+  position: absolute;
+  top: 223px;
+  font-size: 24px;
+}
+
+.no-page-image-container .no-page-image {
+  width: 405px;
+  height: 283px;
+  position: relative;
+  left: 75px;
+  background-image: url("../../assets/sniffing-dog.gif");
+  background-repeat: no-repeat;
+  background-size: contain;
+}
+/*# sourceMappingURL=NoPage.css.map */
