+++ src/pages/profile/Profile.css
@@ -1,0 +1,48 @@
+.profile-page-container {
+  padding-bottom: 28px;
+  display: -webkit-box;
+  display: -ms-flexbox;
+  display: flex;
+  -webkit-box-pack: space-evenly;
+      -ms-flex-pack: space-evenly;
+          justify-content: space-evenly;
+  background-color: #FFF2EF;
+}
+
+.profile-page-container::before {
+  content: '';
+  max-width: 100%;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  right: 0;
+  background-color: #FFF2EF;
+  z-index: -1;
+}
+
+.profile-page-container h2 {
+  margin-top: 32px;
+  margin-bottom: 32px;
+  font-size: 28px;
+  font-weight: 500;
+}
+
+.profile-page-container .profile-right {
+  margin-bottom: 40px;
+}
+
+.profile-page-container .profile-right .profile-sent-names-container {
+  max-width: 521px;
+}
+
+.profile-page-container .dog-card-large-background {
+  background-color: rgba(255, 242, 239, 0.88);
+  width: 100vw;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 99;
+}
+/*# sourceMappingURL=Profile.css.map */
