+++ src/pages/profile/Profile.scss
@@ -1,0 +1,43 @@
+.profile-page-container {
+    padding-bottom: 28px;
+    display: flex;
+    justify-content: space-evenly;
+    background-color: #FFF2EF;
+    
+    &::before {
+        content: '';
+        max-width: 100%;
+        height: 100%;
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        background-color:  #FFF2EF;
+        z-index: -1;
+    }
+    
+    h2 {
+        margin-top: 32px;
+        margin-bottom: 32px;
+        font-size: 28px;
+        font-weight: 500;
+    }
+
+    .profile-right {
+        margin-bottom: 40px;
+
+        .profile-sent-names-container {
+            max-width: 521px;
+        }
+    }
+
+    .dog-card-large-background {
+        background-color: rgba(255, 242, 239, 0.88);
+        width: 100vw;
+        height: 100vh;
+        position: fixed;
+        top: 0;
+        left: 0;
+        z-index: 99;
+    }
+}
