+++ src/reset.css
@@ -1,34 +1,34 @@
 * {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  outline: none;
-  box-sizing: border-box;
-  background: transparent;
-  font-family: inherit;
+    margin: 0;
+    padding: 0;
+    border: 0;
+    outline: none;
+    box-sizing: border-box;
+    background: transparent;
+   font-family: inherit;
 }
 
 a {
-  text-decoration: none;
+    text-decoration: none;
 }
 
 a img {
-  border: none;
+    border: none;
 }
 
 button,
 a {
-  cursor: pointer;
-  color: inherit;
+    cursor: pointer;
+    color: inherit;
 }
 
 textarea {
-  resize: none;
+    resize: none;
 }
 
 button,
 textarea,
 select {
-  appearance: none !important;
-  border-radius: 0;
+    appearance: none !important;
+    border-radius: 0;
 }
